Manual for percona database xtrabackup
The PDF contains the manual for mysql percona xtrabackup. It contains the script and explanation for both backup and restore as well as the automation. The backup can be done with a single schema or a single table or the whole database. It also contains the code for the backup automation.